chore: refactorizacion de carpetas y archivos de sln
