+++ .dockerignore
@@ -3,6 +3,7 @@
 **/.env
 **/.git
 **/.gitignore
+tests/
 **/.project
 **/.settings
 **/.toolstarget
