+++ AcadEvalSys.sln
@@ -2,9 +2,9 @@
 # Visual Studio Version 17
 VisualStudioVersion = 17.10.34928.147
 MinimumVisualStudioVersion = 10.0.40219.1
-Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "AcadEvalSys.WEB.Client", "src\AcadEval.WEB\AcadEvalSys.WEB.Client\AcadEvalSys.WEB.Client.esproj", "{E78FF921-1CA7-4DA3-A871-1991D4B300A0}"
+Project("{54A90642-561A-4BB1-A94E-469ADEE60C69}") = "AcadEvalSys.WEB.Client", "src\AcadEvalSys.WEB\AcadEvalSys.WEB.Client\AcadEvalSys.WEB.Client.esproj", "{E78FF921-1CA7-4DA3-A871-1991D4B300A0}"
 EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcadEvalSys.API", "src\AcadEval.WEB\AcadEval.WEB.Server\AcadEvalSys.API.csproj", "{E13D0CBD-7865-486B-BB56-7D38FBAAB745}"
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcadEvalSys.API", "src\AcadEvalSys.WEB\AcadEvalSys.WEB.Server\AcadEvalSys.API.csproj", "{E13D0CBD-7865-486B-BB56-7D38FBAAB745}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "AcadEvalSys.Domain", "src\AcadEvalSys.Domain\AcadEvalSys.Domain.csproj", "{C5B6BCCC-675F-4BDE-B676-9AC7B05993B3}"
 EndProject
