--- src/AcadEval.WEB/AcadEvalSys.WEB.Client/vite.config.ts
@@ -1,88 +1,0 @@
-import {fileURLToPath, URL} from "node:url";
-import {defineConfig} from "vite";
-import react from "@vitejs/plugin-react";
-import tailwindcss from "@tailwindcss/vite";
-import fs from "fs";
-import path from "path";
-import child_process from "child_process";
-import {env} from "process";
-
-const baseFolder =
-    env.APPDATA !== undefined && env.APPDATA !== ""
-        ? `${env.APPDATA}/ASP.NET/https`
-        : `${env.HOME}/.aspnet/https`;
-console.log(baseFolder);
-const certificateName = "reacttest.client";
-const certFilePath = path.join(baseFolder, `${certificateName}.pem`);
-const keyFilePath = path.join(baseFolder, `${certificateName}.key`);
-
-if (!fs.existsSync(certFilePath) || !fs.existsSync(keyFilePath)) {
-    if (
-        0 !==
-        child_process.spawnSync(
-            "dotnet",
-            [
-                "dev-certs",
-                "https",
-                "--export-path",
-                certFilePath,
-                "--format",
-                "Pem",
-                "--no-password",
-            ],
-            {stdio: "inherit"}
-        ).status
-    ) {
-        throw new Error("Could not create certificate.");
-    }
-}
-
-// Configura la URL del API basada en variables de entorno o usa valores predeterminados
-const target = env.ASPNETCORE_HTTPS_PORT
-    ? `https://localhost:${env.ASPNETCORE_HTTPS_PORT}`
-    : env.ASPNETCORE_URLS
-        ? env.ASPNETCORE_URLS.split(";")[0]
-        : "https://localhost:7004";
-
-console.log(`API proxy target: ${target}`);
-
-export default defineConfig({
-    plugins: [react(), tailwindcss()],
-    resolve: {
-        alias: {
-            "@": fileURLToPath(new URL("./src", import.meta.url)),
-        },
-    },
-    server: {
-        proxy: {
-            "^/api/": {
-                target,
-                secure: false,
-                changeOrigin: true,
-                rewrite: (path) => path.replace(/^\/api/, '/api'),
-                configure: (proxy, _options) => {
-                    proxy.on('error', (err, _req, _res) => {
-                        console.log('proxy error', err);
-                    });
-                    proxy.on('proxyReq', (proxyReq, req, _res) => {
-                        console.log('Sending Request to the Target:', req.method, req.url);
-                    });
-                    proxy.on('proxyRes', (proxyRes, req, _res) => {
-                        console.log('Received Response from the Target:', proxyRes.statusCode, req.url);
-                    });
-                },
-            },
-        },
-        port: 5173,
-        https: {
-            key: fs.readFileSync(keyFilePath),
-            cert: fs.readFileSync(certFilePath),
-        },
-        // Configura cors para permitir solicitudes desde la API
-        cors: true,
-        // Imprime la URL donde está corriendo el cliente
-        hmr: {
-            clientPort: 5173,
-        },
-    },
-});
