+++ src/AcadEvalSys.Domain/AcadEvalSys.Domain.csproj
@@ -13,9 +13,10 @@
     </ItemGroup>
 
     <ItemGroup>
-      <Reference Include="Microsoft.Extensions.Identity.Stores">
+      <!-- <Reference Include="Microsoft.Extensions.Identity.Stores">
         <HintPath>..\..\..\..\.nuget\packages\microsoft.aspnetcore.app.ref\8.0.15\ref\net8.0\Microsoft.Extensions.Identity.Stores.dll</HintPath>
-      </Reference>
+      </Reference> -->
+      <PackageReference Include="Microsoft.AspNetCore.Identity.EntityFrameworkCore" Version="8.0.7" />
     </ItemGroup>
 
 </Project>
