+++ src/AcadEvalSys.WEB/AcadEvalSys.WEB.Client/nuget.config
